requirements.txt
# Core dependencies (needed for ETL, DB and basic analysis)
pandas>=1.5.0
psycopg2-binary>=2.9.6
SQLAlchemy>=1.4.0
requests>=2.28.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
beautifulsoup4>=4.12.0

# Dev & tooling
pytest>=7.0.0
jupyterlab>=4.0.0
ipykernel>=6.0.0


sentence-transformers>=2.2.2    
torch>=2.0.0                     
flashtext>=2.7                   
git add <resolved-file>
git rebase --continue